Fix announcement_user migration foreign key types

+++ database/migrations/2026_02_12_200300_create_announcement_user_table.php
@@ -14,7 +14,7 @@ class CreateAnnouncementUserTable extends Migration
     public function up()
     {
         Schema::create('announcement_user', function (Blueprint $table) {
-            $table->unsignedBigInteger('announcement_id');
+            $table->unsignedInteger('announcement_id');
             $table->unsignedInteger('user_id');
             $table->timestamp('dismissed_at')->useCurrent();
